Organize utils into various files.
